Split off record batch implementation into separate files, progress refactoring
