+++ cpp/src/arrow/api.h
@@ -27,6 +27,7 @@
 #include "arrow/memory_pool.h"
 #include "arrow/pretty_print.h"
 #include "arrow/status.h"
+#include "arrow/record_batch.h"
 #include "arrow/table.h"
 #include "arrow/table_builder.h"
 #include "arrow/tensor.h"
