+++ cpp/src/arrow/gpu/cuda_arrow_ipc.cc
@@ -28,7 +28,7 @@
 #include "arrow/ipc/reader.h"
 #include "arrow/ipc/writer.h"
 #include "arrow/status.h"
-#include "arrow/table.h"
+#include "arrow/record_batch.h"
 #include "arrow/util/visibility.h"
 
 #include "arrow/gpu/cuda_context.h"
