+++ cpp/src/arrow/pretty_print.cc
@@ -22,8 +22,8 @@
 
 #include "arrow/array.h"
 #include "arrow/pretty_print.h"
+#include "arrow/record_batch.h"
 #include "arrow/status.h"
-#include "arrow/table.h"
 #include "arrow/type.h"
 #include "arrow/type_traits.h"
 #include "arrow/util/logging.h"
