+++ cpp/src/arrow/table.h
@@ -24,6 +24,7 @@
 #include <vector>
 
 #include "arrow/array.h"
+#include "arrow/record_batch.h"
 #include "arrow/type.h"
 #include "arrow/util/macros.h"
 #include "arrow/util/visibility.h"
@@ -32,8 +33,6 @@ namespace arrow {
 
 class KeyValueMetadata;
 class Status;
-
-using ArrayVector = std::vector<std::shared_ptr<Array>>;
 
 /// \class ChunkedArray
 /// \brief A data structure managing a list of primitive Arrow arrays logically
@@ -113,114 +112,6 @@ class ARROW_EXPORT Column {
   ARROW_DISALLOW_COPY_AND_ASSIGN(Column);
 };
 
-/// \class RecordBatch
-/// \brief Collection of equal-length arrays matching a particular Schema
-///
-/// A record batch is table-like data structure that is semantically a sequence
-/// of fields, each a contiguous Arrow array
-class ARROW_EXPORT RecordBatch {
- public:
-  virtual ~RecordBatch() = default;
-
-  /// \param[in] schema The record batch schema
-  /// \param[in] num_rows length of fields in the record batch. Each array
-  /// should have the same length as num_rows
-  /// \param[in] columns the record batch fields as vector of arrays
-  static std::shared_ptr<RecordBatch> Make(
-      const std::shared_ptr<Schema>& schema,
-      int64_t num_rows, const std::vector<std::shared_ptr<Array>>& columns);
-
-  /// \brief Move-based constructor for a vector of Array instances
-  static std::shared_ptr<RecordBatch> Make(
-      const std::shared_ptr<Schema>& schema, int64_t num_rows,
-      std::vector<std::shared_ptr<Array>>&& columns);
-
-  /// \brief Construct record batch from vector of internal data structures
-  /// \since 0.5.0
-  ///
-  /// This class is only provided with an rvalue-reference for the input data,
-  /// and is intended for internal use, or advanced users.
-  ///
-  /// \param schema the record batch schema
-  /// \param num_rows the number of semantic rows in the record batch. This
-  /// should be equal to the length of each field
-  /// \param columns the data for the batch's columns
-  static std::shared_ptr<RecordBatch> Make(
-      const std::shared_ptr<Schema>& schema, int64_t num_rows,
-      std::vector<std::shared_ptr<ArrayData>>&& columns);
-
-  /// \brief Construct record batch by copying vector of array data
-  /// \since 0.5.0
-  static std::shared_ptr<RecordBatch> Make(
-      const std::shared_ptr<Schema>& schema, int64_t num_rows,
-      const std::vector<std::shared_ptr<ArrayData>>& columns);
-
-  /// \brief Determine if two record batches are exactly equal
-  /// \return true if batches are equal
-  bool Equals(const RecordBatch& other) const;
-
-  /// \brief Determine if two record batches are approximately equal
-  bool ApproxEquals(const RecordBatch& other) const;
-
-  // \return the table's schema
-  /// \return true if batches are equal
-  std::shared_ptr<Schema> schema() const = 0;
-
-  /// \brief Retrieve an array from the record batch
-  /// \param[in] i field index, does not boundscheck
-  /// \return an Array object
-  virtual std::shared_ptr<Array> column(int i) const = 0;
-
-  std::shared_ptr<ArrayData> column_data(int i) const = 0;
-
-  /// \brief Name in i-th column
-  const std::string& column_name(int i) const;
-
-  /// \return the number of columns in the table
-  int num_columns() const = 0;
-
-  // { return static_cast<int>(columns_.size()); }
-
-  /// \return the number of rows (the corresponding length of each column)
-  int64_t num_rows() const { return num_rows_; }
-
-  /// \brief Replace schema key-value metadata with new metadata (EXPERIMENTAL)
-  /// \since 0.5.0
-  ///
-  /// \param[in] metadata new KeyValueMetadata
-  /// \return new RecordBatch
-  std::shared_ptr<RecordBatch> ReplaceSchemaMetadata(
-      const std::shared_ptr<const KeyValueMetadata>& metadata) const;
-
-  /// \brief Slice each of the arrays in the record batch
-  /// \param[in] offset the starting offset to slice, through end of batch
-  /// \return new record batch
-  virtual std::shared_ptr<RecordBatch> Slice(int64_t offset) const = 0;
-
-  /// \brief Slice each of the arrays in the record batch
-  /// \param[in] offset the starting offset to slice
-  /// \param[in] length the number of elements to slice from offset
-  /// \return new record batch
-  virtual std::shared_ptr<RecordBatch> Slice(int64_t offset,
-                                             int64_t length) const = 0;
-
-  /// \brief Check for schema or length inconsistencies
-  /// \return Status
-  Status Validate() const;
-
- private:
-  ARROW_DISALLOW_COPY_AND_ASSIGN(RecordBatch);
-
-  RecordBatch(const std::shared_ptr<Schema>& schema, int64_t num_rows);
-
-  std::shared_ptr<Schema> schema_;
-  int64_t num_rows_;
-  std::vector<std::shared_ptr<ArrayData>> columns_;
-
-  // Caching boxed array data
-  mutable std::vector<std::shared_ptr<Array>> boxed_columns_;
-};
-
 /// \class Table
 /// \brief Logical table as sequence of chunked arrays
 class ARROW_EXPORT Table {
@@ -290,22 +181,6 @@ class ARROW_EXPORT Table {
   std::vector<std::shared_ptr<Column>> columns_;
 
   int64_t num_rows_;
-};
-
-/// \brief Abstract interface for reading stream of record batches
-class ARROW_EXPORT RecordBatchReader {
- public:
-  virtual ~RecordBatchReader();
-
-  /// \return the shared schema of the record batches in the stream
-  virtual std::shared_ptr<Schema> schema() const = 0;
-
-  /// Read the next record batch in the stream. Return null for batch when
-  /// reaching end of stream
-  ///
-  /// \param[out] batch the next loaded batch, null at end of stream
-  /// \return Status
-  virtual Status ReadNext(std::shared_ptr<RecordBatch>* batch) = 0;
 };
 
 /// \brief Compute a sequence of record batches from a (possibly chunked) Table
