+++ cpp/src/arrow/test-common.h
@@ -30,7 +30,6 @@
 #include "arrow/buffer.h"
 #include "arrow/builder.h"
 #include "arrow/memory_pool.h"
-#include "arrow/table.h"
 #include "arrow/test-util.h"
 
 namespace arrow {
